Move tailer into batchers package
